외형 수량 자료형 불일치로 인한 오류 수정

+++ src/main/java/com/bumsoap/store/dto/ShapeSelItem.java
@@ -9,6 +9,8 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 public class ShapeSelItem {
   private String shapeLabel;
-  private int count;
+
+  private Long count;
+  
   private BigDecimal price;
 }
